requirements.txt
Flask
Flask-SocketIO
pydub
numpy
python-dotenv
